requirements.txt
strawberry-graphql[debug-server]
uvicorn
starlette
redis
pyotp
pyjwt
pydantic[email]
pytest